@@ -2,34 +2,38 @@
   "name": "fk-atem-control-service",
   "version": "1.0.0",
   "description": "Frikanalen ATEM control microservice",
-  "main": "index.js",
+  "exports": "./src/index.js",
   "dependencies": {
-    "@tsconfig/node14": "^1.0.0",
-    "atem-connection": "^1.3.0",
-    "body-parser": "^1.19.0",
-    "cookie-parser": "^1.4.5",
-    "express": "^4.17.1",
-    "node-fetch": "^2.6.7",
+    "atem-connection": "^3.1.0",
+    "body-parser": "^1.20.1",
+    "cookie-parser": "^1.4.6",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "node-fetch": "^3.3.0",
     "start": "^5.1.0",
-    "winston": "^3.3.3"
+    "winston": "^3.8.2",
+    "zod": "^3.19.1"
   },
   "devDependencies": {
-    "@tsconfig/node16-strictest": "^1.0.1",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/express": "^4.17.9",
-    "@types/express-serve-static-core": "^4.17.17",
-    "@types/node": "^14.14.14",
-    "@types/node-fetch": "^2.5.7",
-    "prettier": "^2.6.2",
+    "@tsconfig/node18-strictest-esm": "^1",
+    "@types/cookie-parser": "^1.4.3",
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.14",
+    "@types/express-serve-static-core": "^4.17.31",
+    "@types/node": "^18.11.9",
+    "@types/node-fetch": "^2.6.2",
+    "prettier": "^2.7.1",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^4.7.3"
+    "typescript": "^4.9.3"
   },
   "scripts": {
     "build": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node dist/server.js",
-    "dev": "ts-node-dev src/server.ts"
+    "dev": "NODE_ENV=development ts-node-esm ./src/server.ts"
   },
+  "type": "module",
   "author": "",
   "license": "MIT"
 }
