@@ -1,7 +1,7 @@
 import type { RequestHandler } from "express";
-import atem from "./AtemInterface";
+import atem from "./AtemInterface.js";
 
-export const setInput: RequestHandler = async (req, res) => {
+export const setProgram: RequestHandler = async (req, res) => {
   const inputIndex = parseInt(req.body.inputIndex);
 
   if (isNaN(inputIndex)) {
@@ -13,14 +13,37 @@ export const setInput: RequestHandler = async (req, res) => {
   }
 
   try {
-    await atem.ME[0]!.setInput(inputIndex);
+    await atem.ME[0]!.setProgram(inputIndex);
   } catch (e: any) {
     res.status(500).send(`could not set input; e = ${e}`);
     return;
   }
-  res.send({ inputIndex: atem.ME[0]?.input });
+  res.send({ inputIndex: atem.ME[0]?.program });
 };
 
-export const getInput: RequestHandler = (_req, res) => {
-  res.send({ inputIndex: atem.ME[0]!.input });
+export const setPreview: RequestHandler = async (req, res) => {
+  const inputIndex = parseInt(req.body.inputIndex);
+
+  if (isNaN(inputIndex)) {
+    res
+      .status(400)
+      .send("Malformed request, expected {inputIndex: number}")
+      .end();
+    return;
+  }
+
+  try {
+    await atem.ME[0]!.setPreview(inputIndex);
+  } catch (e: any) {
+    res.status(500).send(`could not set input; e = ${e}`);
+    return;
+  }
+  res.send({ inputIndex: atem.ME[0]?.program });
+};
+
+export const getProgram: RequestHandler = (_req, res) => {
+  res.send({ inputIndex: atem.ME[0]!.program });
+};
+export const getPreview: RequestHandler = (_req, res) => {
+  res.send({ inputIndex: atem.ME[0]!.preview });
 };
