@@ -1,6 +1,6 @@
 import type { Atem } from "atem-connection";
-import { AudioMixOption } from "atem-connection/dist/enums";
-import { MULTI_VIEWER_INPUT } from "../server";
+import { AudioMixOption } from "atem-connection/dist/enums/index.js";
+import { MULTI_VIEWER_INPUT } from "../server.js";
 
 /**
  * Apply initial state to the ATEM mixer.
@@ -10,19 +10,17 @@ import { MULTI_VIEWER_INPUT } from "../server";
 export const applyInitialConfiguration = async (atem: Atem) => {
   const afv = AudioMixOption.AudioFollowVideo;
 
-    console.log("HIII")
-    await atem.setMultiViewerSource({source: 1, windowIndex: 2});
-    await atem.setMultiViewerSource({source: 2, windowIndex: 3});
-    await atem.setMultiViewerSource({source: 3, windowIndex: 4});
-    await atem.setMultiViewerSource({source: 4, windowIndex: 5});
-    await atem.setMultiViewerSource({source: 5, windowIndex: 6});
+  await atem.setMultiViewerWindowSource(2, 0, 3);
+  await atem.setMultiViewerWindowSource(1, 0, 2);
+  await atem.setMultiViewerWindowSource(3, 0, 4);
+  await atem.setMultiViewerWindowSource(4, 0, 5);
+  await atem.setMultiViewerWindowSource(5, 0, 6);
   try {
-    await atem.setAudioMixerInputProps(1, { mixOption: afv });
-    await atem.setAudioMixerInputProps(2, { mixOption: afv });
-    await atem.setAudioMixerInputProps(3, { mixOption: afv });
-    await atem.setAudioMixerInputProps(4, { mixOption: afv });
-    await atem.setAudioMixerInputProps(5, { mixOption: afv });
-
+    await atem.setClassicAudioMixerInputProps(1, { mixOption: afv });
+    await atem.setClassicAudioMixerInputProps(2, { mixOption: afv });
+    await atem.setClassicAudioMixerInputProps(3, { mixOption: afv });
+    await atem.setClassicAudioMixerInputProps(4, { mixOption: afv });
+    await atem.setClassicAudioMixerInputProps(5, { mixOption: afv });
   } catch (e) {
     console.log("got exception while setting audio mixer input props");
     console.log(e);
