@@ -1,5 +1,5 @@
 import type { RequestHandler } from "express";
-import { checkIfStaff } from "../auth";
+import { checkIfStaff } from "../auth.js";
 
 export const checkStaff: RequestHandler = async (req, res, next) => {
   if (!(await checkIfStaff(req.cookies))) {
