@@ -1,5 +1,5 @@
-import { posterPreview, posterUpload } from "./handlers";
-import { checkStaff } from "../auth/checkStaff";
+import { posterPreview, posterUpload } from "./handlers.js";
+import { checkStaff } from "../auth/checkStaff.js";
 import express from "express";
 
 export const posterRouter = express.Router();
